requirements.txt
# Core Streamlit Web App
streamlit>=1.32.0

# PDF Parsing
PyMuPDF==1.23.4  # for fast, reliable PDF text extraction

# Google Gemini LLM (official SDK)
google-generativeai>=0.5.0

# Optional: Webcam support for proctoring features in test tab
streamlit-webrtc>=0.47.4

# Environment Variables (optional for API key storage)
python-dotenv>=1.0.0

# For working with JSON, rich markdown display, etc.
markdown